+++ formula-android-tests/src/main/java/com/instacart/formula/test/TestLifecycleKey.kt
@@ -1,0 +1,9 @@
+package com.instacart.formula.test
+
+import com.instacart.formula.android.FragmentKey
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class TestLifecycleKey(
+    override val tag: String = "task list",
+) : FragmentKey