+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentAndroidEventTest.kt
@@ -4,9 +4,11 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
+import com.instacart.formula.android.Feature
+import com.instacart.formula.android.ViewFactory
 import com.instacart.formula.android.events.ActivityResult
 import com.instacart.formula.test.TestFragmentActivity
-import com.instacart.formula.test.TestLifecycleContract
+import com.instacart.formula.test.TestLifecycleKey
 import io.reactivex.rxjava3.core.Observable
 import org.junit.Before
 import org.junit.Rule
@@ -24,14 +26,20 @@ class FragmentAndroidEventTest {
                 activity<TestFragmentActivity> {
                     store(
                         configureActivity = {
-                            initialContract = TestLifecycleContract()
+                            initialContract = TestLifecycleKey()
                         },
                         contracts =  {
-                            bind(TestLifecycleContract::class) { _ ->
-                                activityResults().flatMap {
-                                    activityResults.add(it)
-                                    Observable.empty<Any>()
-                                }
+
+                            bind<TestLifecycleKey> { _, _ ->
+                                Feature(
+                                    state = activityResults().flatMap {
+                                        activityResults.add(it)
+                                        Observable.empty()
+                                    },
+                                    viewFactory = ViewFactory.fromLayout(R.layout.test_empty_layout) {
+                                        featureView { }
+                                    }
+                                )
                             }
                         }
                     )