+++ formula-android/src/test/java/com/instacart/formula/android/fakes/TestSignUpFragmentKey.kt
@@ -4,7 +4,6 @@ import com.instacart.formula.android.FragmentKey
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class DetailKey(
-    val id: Int,
-    override val tag: String = "detail-$id",
+data class TestSignUpFragmentKey(
+    override val tag: String = "sign up fragment",
 ) : FragmentKey