+++ formula/src/main/java/com/instacart/formula/internal/FormulaManagerImpl.kt
@@ -3,6 +3,7 @@ package com.instacart.formula.internal
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Formula
 import com.instacart.formula.IFormula
+import com.instacart.formula.Snapshot
 import com.instacart.formula.Transition
 
 /**
@@ -17,28 +18,18 @@ import com.instacart.formula.Transition
 internal class FormulaManagerImpl<Input, State, Output>(
     private val formula: Formula<Input, State, Output>,
     initialInput: Input,
-    private val listeners: ScopedListeners,
-    private val transitionListener: TransitionListener
-) : SnapshotImpl.Delegate, FormulaManager<Input, Output> {
-
-    constructor(
-        formula: Formula<Input, State, Output>,
-        input: Input,
-        transitionListener: TransitionListener
-    ): this(formula, input, ScopedListeners(formula), transitionListener)
-
-    private val actionManager = ActionManager()
-
-    private var children: SingleRequestMap<Any, FormulaManager<*, *>>? = null
-    private var frame: Frame<Input, State, Output>? = null
-    var terminated = false
+    private val transitionListener: TransitionListener,
+    private val listeners: Listeners = Listeners(),
+    private val actionManager: ActionManager = ActionManager(),
+) : FormulaManager<Input, Output> {
 
     private var state: State = formula.initialState(initialInput)
-    private var pendingRemoval: MutableList<FormulaManager<*, *>>? = null
+    private var frame: Frame<Input, State, Output>? = null
+    private var childrenManager: ChildrenManager? = null
 
-    private var childTransitionListener: TransitionListener? = null
+    var terminated = false
 
-    private fun handleTransitionResult(result: Transition.Result<State>) {
+    fun handleTransitionResult(result: Transition.Result<State>) {
         if (terminated) {
             // State transitions are ignored, only side effects are passed up to be executed.
             transitionListener.onTransitionResult(result, true)
@@ -56,9 +47,9 @@ internal class FormulaManagerImpl<Input, State, Output>(
 
     override fun updateTransitionId(transitionId: TransitionId) {
         val lastFrame = checkNotNull(frame) { "missing frame means this is called before initial evaluate" }
-        lastFrame.transitionDispatcher.transitionId = transitionId
+        lastFrame.snapshot.transitionId = transitionId
 
-        children?.forEachValue { it.updateTransitionId(transitionId) }
+        childrenManager?.updateTransitionId(transitionId)
     }
 
     /**
@@ -80,35 +71,21 @@ internal class FormulaManagerImpl<Input, State, Output>(
             state = formula.onInputChanged(prevInput, input, state)
         }
 
-        listeners.evaluationStarted()
-        val transitionDispatcher = TransitionDispatcher(input, state, this::handleTransitionResult, transitionId)
-        val snapshot = SnapshotImpl(transitionId, listeners, this, transitionDispatcher)
-        val result = snapshot.run { formula.run { evaluate() } }
-        val frame = Frame(input, state, result, transitionDispatcher)
+        val snapshot = SnapshotImpl(input, state, transitionId, listeners, this)
+        val result = formula.evaluate(snapshot)
+        val frame = Frame(snapshot, result)
         actionManager.updateEventListeners(frame.evaluation.actions)
         this.frame = frame
 
         listeners.evaluationFinished()
+        childrenManager?.evaluationFinished()
 
-        children?.clearUnrequested {
-            pendingRemoval = pendingRemoval ?: mutableListOf()
-            it.markAsTerminated()
-            pendingRemoval?.add(it)
-        }
-
-        transitionDispatcher.running = true
+        snapshot.running = true
         return result
     }
 
     override fun terminateDetachedChildren(transitionId: TransitionId): Boolean {
-        val local = pendingRemoval
-        pendingRemoval = null
-        local?.forEach { it.performTerminationSideEffects() }
-        if (transitionId.hasTransitioned()) {
-            return true
-        }
-
-        return children?.any { it.value.value.terminateDetachedChildren(transitionId) } ?: false
+        return childrenManager?.terminateDetachedChildren(transitionId) == true
     }
 
     // TODO: should probably terminate children streams, then self.
@@ -120,10 +97,8 @@ internal class FormulaManagerImpl<Input, State, Output>(
         }
 
         // Step through children frames
-        children?.forEachValue {
-            if (it.terminateOldUpdates(transitionId)) {
-                return true
-            }
+        if (childrenManager?.terminateOldUpdates(transitionId) == true) {
+            return true
         }
 
         return false
@@ -138,75 +113,56 @@ internal class FormulaManagerImpl<Input, State, Output>(
         }
 
         // Step through children frames
-        children?.forEachValue {
-            if (it.startNewUpdates(transitionId)) {
-                return true
-            }
+        if (childrenManager?.startNewUpdates(transitionId) == true) {
+            return true
         }
 
         return false
     }
 
-    override fun <ChildInput, ChildOutput> child(
+    fun <ChildInput, ChildOutput> child(
+        key: Any,
         formula: IFormula<ChildInput, ChildOutput>,
         input: ChildInput,
         transitionId: TransitionId
     ): ChildOutput {
-        @Suppress("UNCHECKED_CAST")
-        val children = children ?: run {
-            val initialized: SingleRequestMap<Any, FormulaManager<*, *>> = mutableMapOf()
-            this.children = initialized
-            initialized
-        }
-
-        val compositeKey = constructKey(formula, input)
-        val manager = children
-            .findOrInit(compositeKey) {
-                val childTransitionListener = getOrInitChildTransitionListener()
-                val implementation = formula.implementation()
-                FormulaManagerImpl(implementation, input, childTransitionListener)
-            }
-            .requestAccess {
-                throw IllegalStateException("There already is a child with same key: $compositeKey. Override [Formula.key] function.")
-            } as FormulaManager<ChildInput, ChildOutput>
-
+        val childrenManager = getOrInitChildrenManager()
+        val manager = childrenManager.findOrInitChild(key, formula, input)
         return manager.evaluate(input, transitionId).output
     }
 
     override fun markAsTerminated() {
         terminated = true
-        frame?.transitionDispatcher?.terminated = true
-        children?.forEachValue { it.markAsTerminated() }
+        frame?.snapshot?.terminated = true
+        childrenManager?.markAsTerminated()
     }
 
     override fun performTerminationSideEffects() {
-        children?.forEachValue { it.performTerminationSideEffects() }
+        childrenManager?.performTerminationSideEffects()
         actionManager.terminate()
         listeners.disableAll()
     }
 
-    private fun <ChildInput, ChildOutput> constructKey(
-        formula: IFormula<ChildInput, ChildOutput>,
-        input: ChildInput
-    ): Any {
-        return FormulaKey(
-            type = formula.type(),
-            key = formula.key(input)
-        )
-    }
-
-    private fun getOrInitChildTransitionListener(): TransitionListener {
-        return childTransitionListener ?: run {
-            TransitionListener { result, isChildValid ->
+    private fun getOrInitChildrenManager(): ChildrenManager {
+        return childrenManager ?: run {
+            val listener = TransitionListener { result, isChildValid ->
                 val frame = this.frame
                 if (!isChildValid) {
                     frame?.childInvalidated()
                 }
                 val isValid = frame != null && frame.isValid()
                 transitionListener.onTransitionResult(result, isValid)
-            }.apply {
-                childTransitionListener = this
             }
+
+            val value = ChildrenManager(listener)
+            childrenManager = value
+            value
         }
+    }
+
+    private fun Formula<Input, State, Output>.evaluate(
+        snapshot: Snapshot<Input, State>
+    ): Evaluation<Output> {
+        return snapshot.run { evaluate() }
     }
 }