+++ formula/src/main/java/com/instacart/formula/internal/IndexedKey.kt
@@ -1,0 +1,6 @@
+package com.instacart.formula.internal
+
+internal data class IndexedKey(
+    val key: Any,
+    val index: Int
+)