+++ formula/src/test/java/com/instacart/formula/subjects/EventCallbackFormula.kt
@@ -17,7 +17,7 @@ class EventCallbackFormula : Formula<Unit, String, EventCallbackFormula.Output>(
         return Evaluation(
             output = Output(
                 state = state,
-                changeState = context.onEvent<String> { newState ->
+                changeState = context.onEvent { newState ->
                     transition(newState)
                 }
             )